Begin refactor of core code to external library
