--- custom_components/aprilaire/crc.py
@@ -1,271 +1,0 @@
-"""Utils for computing CRC"""
-
-from __future__ import annotations
-
-CRC_MAP = {
-    0: 0,
-    1: 49,
-    2: 98,
-    3: 83,
-    4: 196,
-    5: 245,
-    6: 166,
-    7: 151,
-    8: 185,
-    9: 136,
-    10: 219,
-    11: 234,
-    12: 125,
-    13: 76,
-    14: 31,
-    15: 46,
-    16: 67,
-    17: 114,
-    18: 33,
-    19: 16,
-    20: 135,
-    21: 182,
-    22: 229,
-    23: 212,
-    24: 250,
-    25: 203,
-    26: 152,
-    27: 169,
-    28: 62,
-    29: 15,
-    30: 92,
-    31: 109,
-    32: 134,
-    33: 183,
-    34: 228,
-    35: 213,
-    36: 66,
-    37: 115,
-    38: 32,
-    39: 17,
-    40: 63,
-    41: 14,
-    42: 93,
-    43: 108,
-    44: 251,
-    45: 202,
-    46: 153,
-    47: 168,
-    48: 197,
-    49: 244,
-    50: 167,
-    51: 150,
-    52: 1,
-    53: 48,
-    54: 99,
-    55: 82,
-    56: 124,
-    57: 77,
-    58: 30,
-    59: 47,
-    60: 184,
-    61: 137,
-    62: 218,
-    63: 235,
-    64: 61,
-    65: 12,
-    66: 95,
-    67: 110,
-    68: 249,
-    69: 200,
-    70: 155,
-    71: 170,
-    72: 132,
-    73: 181,
-    74: 230,
-    75: 215,
-    76: 64,
-    77: 113,
-    78: 34,
-    79: 19,
-    80: 126,
-    81: 79,
-    82: 28,
-    83: 45,
-    84: 186,
-    85: 139,
-    86: 216,
-    87: 233,
-    88: 199,
-    89: 246,
-    90: 165,
-    91: 148,
-    92: 3,
-    93: 50,
-    94: 97,
-    95: 80,
-    96: 187,
-    97: 138,
-    98: 217,
-    99: 232,
-    100: 127,
-    101: 78,
-    102: 29,
-    103: 44,
-    104: 2,
-    105: 51,
-    106: 96,
-    107: 81,
-    108: 198,
-    109: 247,
-    110: 164,
-    111: 149,
-    112: 248,
-    113: 201,
-    114: 154,
-    115: 171,
-    116: 60,
-    117: 13,
-    118: 94,
-    119: 111,
-    120: 65,
-    121: 112,
-    122: 35,
-    123: 18,
-    124: 133,
-    125: 180,
-    126: 231,
-    127: 214,
-    128: 122,
-    129: 75,
-    130: 24,
-    131: 41,
-    132: 190,
-    133: 143,
-    134: 220,
-    135: 237,
-    136: 195,
-    137: 242,
-    138: 161,
-    139: 144,
-    140: 7,
-    141: 54,
-    142: 101,
-    143: 84,
-    144: 57,
-    145: 8,
-    146: 91,
-    147: 106,
-    148: 253,
-    149: 204,
-    150: 159,
-    151: 174,
-    152: 128,
-    153: 177,
-    154: 226,
-    155: 211,
-    156: 68,
-    157: 117,
-    158: 38,
-    159: 23,
-    160: 252,
-    161: 205,
-    162: 158,
-    163: 175,
-    164: 56,
-    165: 9,
-    166: 90,
-    167: 107,
-    168: 69,
-    169: 116,
-    170: 39,
-    171: 22,
-    172: 129,
-    173: 176,
-    174: 227,
-    175: 210,
-    176: 191,
-    177: 142,
-    178: 221,
-    179: 236,
-    180: 123,
-    181: 74,
-    182: 25,
-    183: 40,
-    184: 6,
-    185: 55,
-    186: 100,
-    187: 85,
-    188: 194,
-    189: 243,
-    190: 160,
-    191: 145,
-    192: 71,
-    193: 118,
-    194: 37,
-    195: 20,
-    196: 131,
-    197: 178,
-    198: 225,
-    199: 208,
-    200: 254,
-    201: 207,
-    202: 156,
-    203: 173,
-    204: 58,
-    205: 11,
-    206: 88,
-    207: 105,
-    208: 4,
-    209: 53,
-    210: 102,
-    211: 87,
-    212: 192,
-    213: 241,
-    214: 162,
-    215: 147,
-    216: 189,
-    217: 140,
-    218: 223,
-    219: 238,
-    220: 121,
-    221: 72,
-    222: 27,
-    223: 42,
-    224: 193,
-    225: 240,
-    226: 163,
-    227: 146,
-    228: 5,
-    229: 52,
-    230: 103,
-    231: 86,
-    232: 120,
-    233: 73,
-    234: 26,
-    235: 43,
-    236: 188,
-    237: 141,
-    238: 222,
-    239: 239,
-    240: 130,
-    241: 179,
-    242: 224,
-    243: 209,
-    244: 70,
-    245: 119,
-    246: 36,
-    247: 21,
-    248: 59,
-    249: 10,
-    250: 89,
-    251: 104,
-    252: 255,
-    253: 206,
-    254: 157,
-    255: 172,
-}
-
-
-def generate_crc(data: list[int]) -> int:
-    """Generate the CRC for a data payload"""
-    crc = int(0)
-    for i in data:
-        key = i ^ crc
-        crc = CRC_MAP[key]
-    return crc
